+++ pscompose/api/routers/schedules.py
@@ -5,7 +5,7 @@ from fastapi.responses import JSONResponse
 from pscompose.settings import DataTypes
 from pscompose.utils import generate_router
 from pscompose.backends.postgres import backend
-from pscompose.form_schemas import SCHEDULE_SCHEMA, SCHEDULE_UI_SCHEMA
+from pscompose.form_schemas.schedule_schemas import SCHEDULE_SCHEMA, SCHEDULE_UI_SCHEMA
 
 # Setup CRUD endpoints
 router = generate_router("schedule")
