+++ pscompose/form_schemas/test_schemas.py
@@ -1,507 +1,3 @@
-# For static forms that need to be rendered without talking to pScheduler
-# Each DataType will have it's own schema and UI Schema
-
-# TODO: Should the properties be named as how they appear in the config JSON?
-# Eg: using display-name instead of name
-ADDRESS_SCHEMA = {
-    "title": "Schema for creating a new address",
-    "type": "object",
-    "properties": {
-        "name": {
-            "type": "string",
-            "title": "Host Name",
-            "description": "A string to identify this host",
-        },
-        "address": {
-            "type": "string",
-            "title": "Address",
-            "description": "The host address",
-        },
-        "no-agent": {
-            "type": "boolean",
-            "title": "No Agent",
-            "description": "Check this box if no agent is required",
-            "default": False,
-        },
-        "disabled": {
-            "type": "boolean",
-            "title": "Disabled",
-            "description": "Check this box if address is disabled. Set to False by default",
-            "default": False,
-        },
-        "lead-bind-address": {
-            "type": "string",
-            "title": "Lead Bind Address",
-            "description": "This property must be an IP or hostname",
-        },
-        "pscheduler-address": {
-            "type": "string",
-            "title": "pScheduler Address",
-            "description": "This property is an IP or hostname with an optional port specification",
-        },
-        "contexts": {
-            "type": "array",
-            "title": "Contexts",
-            "items": {"oneOf": []},
-        },
-        "_meta": {
-            "type": "object",
-            "title": "Other Meta",
-            "description": "Fill in information such as display-name and display-set as an object",
-        },
-    },
-    "required": [
-        "name",
-        "address",
-    ],
-    "renderers": {},
-}
-
-ADDRESS_UI_SCHEMA = {
-    "type": "VerticalLayout",
-    "elements": [
-        {
-            "type": "HorizontalLayout",
-            "elements": [
-                {
-                    "type": "Control",
-                    "scope": "#/properties/name",
-                    "customComponent": "ps-input-text",
-                },
-                {
-                    "type": "Control",
-                    "scope": "#/properties/disabled",
-                    "customComponent": "ps-input-checkbox",
-                },
-            ],
-        },
-        {
-            "type": "HorizontalLayout",
-            "elements": [
-                {
-                    "type": "Control",
-                    "scope": "#/properties/address",
-                    "customComponent": "ps-input-text",
-                },
-                {
-                    "type": "Control",
-                    "scope": "#/properties/no-agent",
-                    "customComponent": "ps-input-checkbox",
-                },
-            ],
-        },
-        {
-            "type": "Control",
-            "scope": "#/properties/lead-bind-address",
-            "customComponent": "ps-input-text-autocomplete",
-        },
-        {
-            "type": "Control",
-            "scope": "#/properties/pscheduler-address",
-            "customComponent": "ps-input-text-autocomplete",
-        },
-        {
-            "type": "Control",
-            "scope": "#/properties/contexts",
-            "customComponent": "ps-select-multi",
-        },
-        {"type": "Control", "scope": "#/properties/_meta", "customComponent": "ps-textarea-json"},
-    ],
-}
-
-GROUP_SCHEMA = {
-    "title": "Schema for creating a new group",
-    "type": "object",
-    "properties": {
-        "name": {"type": "string", "title": "Name"},
-        "type": {
-            "type": "string",
-            "title": "Type",
-            "oneOf": [
-                {"const": "list", "title": "List"},
-                {"const": "disjoint", "title": "Disjoint"},
-                {"const": "mesh", "title": "Mesh"},
-            ],
-        },
-    },
-    "required": ["name", "type"],
-    "allOf": [
-        {
-            "if": {"properties": {"type": {"const": "list"}}},
-            "then": {
-                "properties": {
-                    "type": {"const": "list"},
-                    "addresses": {"type": "array", "title": "Addresses", "items": {"oneOf": []}},
-                    "_meta": {
-                        "type": "object",
-                        "title": "Other Meta",
-                        "description": "Fill in information such as display-name and display-set as an object",
-                    },
-                },
-                "required": ["addresses"],
-            },
-        },
-        {
-            "if": {"properties": {"type": {"const": "disjoint"}}},
-            "then": {
-                "properties": {
-                    "type": {"const": "disjoint"},
-                    "unidirectional": {
-                        "type": "boolean",
-                        "title": "Unidirectional",
-                        "description": "Check this box if it's unidirectional",
-                        "default": False,
-                    },
-                    "a-addresses": {
-                        "type": "array",
-                        "title": "A-Addresses",
-                        "items": {"oneOf": []},
-                    },
-                    "b-addresses": {
-                        "type": "array",
-                        "title": "B-Addresses",
-                        "items": {"oneOf": []},
-                    },
-                    "excludes-self": {
-                        "type": "string",
-                        "title": "Excludes Self",
-                        "oneOf": [
-                            {"const": "host", "title": "Host"},
-                            {"const": "address", "title": "Address"},
-                            {"const": "disabled", "title": "Disabled"},
-                        ],
-                    },
-                    "excludes": {"type": "array", "title": "Excludes", "items": {"oneOf": []}},
-                    "_meta": {
-                        "type": "object",
-                        "title": "Other Meta",
-                        "description": "Fill in information such as display-name and display-set as an object",
-                    },
-                },
-                "required": ["a-addresses", "b-addresses"],
-            },
-        },
-        {
-            "if": {"properties": {"type": {"const": "mesh"}}},
-            "then": {
-                "properties": {
-                    "type": {"const": "mesh"},
-                    "addresses": {"type": "array", "title": "Addresses", "items": {"oneOf": []}},
-                    "excludes-self": {
-                        "type": "string",
-                        "title": "Excludes Self",
-                        "oneOf": [
-                            {"const": "host", "title": "Host"},
-                            {"const": "address", "title": "Address"},
-                            {"const": "disabled", "title": "Disabled"},
-                        ],
-                    },
-                    "excludes": {"type": "array", "title": "Excludes", "items": {"oneOf": []}},
-                    "_meta": {
-                        "type": "object",
-                        "title": "Other Meta",
-                        "description": "Fill in information such as display-name and display-set as an object",
-                    },
-                },
-                "required": ["addresses"],
-            },
-        },
-    ],
-}
-
-GROUP_UI_SCHEMA = {
-    "type": "VerticalLayout",
-    "elements": [
-        {"type": "Control", "scope": "#/properties/name", "customComponent": "ps-input-text"},
-        {
-            "type": "Control",
-            "scope": "#/properties/type",
-            "customComponent": "ps-select",
-        },
-        {
-            "type": "Group",
-            "rule": {
-                "effect": "SHOW",
-                "condition": {"scope": "#/properties/type", "schema": {"const": "list"}},
-            },
-            "elements": [
-                {
-                    "type": "Control",
-                    "scope": "#/properties/addresses",
-                    "options": {"format": "select"},
-                    "customComponent": "ps-select-multi",
-                },
-                {
-                    "type": "Control",
-                    "scope": "#/properties/_meta",
-                    "customComponent": "ps-textarea-json",
-                },
-            ],
-        },
-        {
-            "type": "Group",
-            "rule": {
-                "effect": "SHOW",
-                "condition": {"scope": "#/properties/type", "schema": {"const": "disjoint"}},
-            },
-            "elements": [
-                {
-                    "type": "Control",
-                    "scope": "#/properties/a-addresses",
-                    "customComponent": "ps-select-multi",
-                },
-                {
-                    "type": "Control",
-                    "scope": "#/properties/b-addresses",
-                    "customComponent": "ps-select-multi",
-                },
-                {
-                    "type": "HorizontalLayout",
-                    "elements": [
-                        {
-                            "type": "Control",
-                            "scope": "#/properties/excludes-self",
-                            "customComponent": "ps-select",
-                        },
-                        {
-                            "type": "Control",
-                            "scope": "#/properties/unidirectional",
-                            "customComponent": "ps-input-checkbox",
-                        },
-                    ],
-                },
-                {
-                    "type": "Control",
-                    "scope": "#/properties/excludes",
-                    "customComponent": "ps-select-excludes",
-                },
-                {
-                    "type": "Control",
-                    "scope": "#/properties/_meta",
-                    "customComponent": "ps-textarea-json",
-                },
-            ],
-        },
-        {
-            "type": "Group",
-            "rule": {
-                "effect": "SHOW",
-                "condition": {"scope": "#/properties/type", "schema": {"const": "mesh"}},
-            },
-            "elements": [
-                {
-                    "type": "Control",
-                    "scope": "#/properties/addresses",
-                    "options": {"format": "select"},
-                    "customComponent": "ps-select-multi",
-                },
-                {
-                    "type": "Control",
-                    "scope": "#/properties/excludes-self",
-                    "customComponent": "ps-select",
-                },
-                {
-                    "type": "Control",
-                    "scope": "#/properties/excludes",
-                    "customComponent": "ps-select-excludes",
-                },
-                {
-                    "type": "Control",
-                    "scope": "#/properties/_meta",
-                    "options": {"multi": True},
-                    "customComponent": "ps-textarea-json",
-                },
-            ],
-        },
-    ],
-}
-
-SCHEDULE_SCHEMA = {
-    "title": "Schema for creating a new schedule",
-    "type": "object",
-    "properties": {
-        "name": {
-            "type": "string",
-            "title": "Schedule Name",
-            "description": "A string to identify this schedule",
-        },
-        "slip": {
-            "type": "string",
-            "title": "Slip",
-            "description": "ISO8601 Duration that allows the start of each run to be as much as the Duration later than their ideal scheduled time. Default is 5 minutes (PT5M)",
-            "default": "PT5M",
-        },
-        "sliprand": {
-            "type": "boolean",
-            "title": "Randomize Slip",
-            "description": "Randomly choose a timeslot within the allowed slip instead of choosing earliest available",
-            "default": True,
-        },
-        "repeat": {
-            "type": "string",
-            "title": "Repeat",
-            "description": "ISO8601 Duration that repeats runs at the specified duration.<br>Examples:<br> PT5M (5 minutes)<br> PT10M (10 minutes)",
-        },
-        # "start": {
-        #     "type": "string",
-        #     "title": "Start",
-        #     "description": "",
-        # },
-        # "until": {
-        #     "type": "string",
-        #     "title": "Until",
-        #     "description": "",
-        # },
-        "max-runs": {
-            "type": "integer",
-            "title": "Max Runs",
-            "description": "Allow the task to run up to N times",
-            "minimum": 1,
-        },
-        "_meta": {
-            "type": "object",
-            "title": "Other Meta",
-            "description": "This field should be a JSON object with keys such as 'display-name' and 'display-set'",
-        },
-    },
-    "required": [
-        "name",
-    ],
-    "renderers": {},
-}
-
-SCHEDULE_UI_SCHEMA = {
-    "type": "VerticalLayout",
-    "elements": [
-        {"type": "Control", "scope": "#/properties/name", "customComponent": "ps-input-text"},
-        {"type": "Control", "scope": "#/properties/repeat", "customComponent": "ps-input-text"},
-        {
-            "type": "HorizontalLayout",
-            "elements": [
-                {
-                    "type": "Control",
-                    "scope": "#/properties/slip",
-                    "customComponent": "ps-input-text",
-                },
-                {
-                    "type": "Control",
-                    "scope": "#/properties/sliprand",
-                    "customComponent": "ps-input-checkbox",
-                },
-            ],
-        },
-        # {
-        #     "type": "Control",
-        #     "scope": "#/properties/start",
-        # },
-        # {
-        #     "type": "Control",
-        #     "scope": "#/properties/until",
-        # },
-        {
-            "type": "Control",
-            "scope": "#/properties/max-runs",
-            "customComponent": "ps-input-number",
-        },
-        {"type": "Control", "scope": "#/properties/_meta", "customComponent": "ps-textarea-json"},
-    ],
-}
-
-ARCHIVE_SCHEMA = {
-    "title": "Schema for creating a new archive",
-    "type": "object",
-    "properties": {
-        "name": {
-            "type": "string",
-            "title": "Archive Name",
-            "description": "A string to identify this archive",
-        },
-        "type": {
-            "type": "string",
-            "title": "Type",
-            "description": "The type of archive",
-            "oneOf": [],
-            # "oneOf": [
-            #     {"const": "bitbucket", "title": "Bitbucket"},
-            #     {"const": "esmond", "title": "ESmond"},
-            #     {"const": "failer", "title": "Failer"},
-            #     {"const": "http", "title": "HTTP"},
-            #     {"const": "kafka", "title": "Kafka"},
-            #     {"const": "postgresql", "title": "PostgreSQL"},
-            #     {"const": "rabbitmq", "title": "RabbitMQ"},
-            #     {"const": "snmptrap", "title": "SNMP Trap"},
-            #     {"const": "syslog", "title": "Syslog"},
-            #     {"const": "tcp", "title": "TCP"},
-            #     {"const": "udp", "title": "UDP"},
-            # ],
-        },
-        # "data": {
-        #     "type": "string",
-        #     "title": "Data",
-        #     "description": "JSON object that specifies archive-specific parameters. Archive objects in pSConfig are taken directly from pScheduler. Eg: _url",
-        # },
-        # "schema": {
-        #     "type": "integer",
-        #     "title": "Schema Version",
-        #     "description": "The schema version number",
-        # },
-        "ttl": {
-            "type": "string",
-            "title": "TTL",
-            "description": "ISO8601 duration that specifies the ttl",
-        },
-        "label": {
-            "type": "string",
-            "title": "Label",
-            "description": "",
-        },
-        "transform": {
-            "type": "string",
-            "title": "Transform",
-            "description": "",
-        },
-        "_meta": {
-            "type": "object",
-            "title": "Other Meta",
-            "description": "Fill in information such as display-name and display-set as an object",
-        },
-    },
-    "required": [
-        "name",
-        "type",
-        # "data",
-    ],
-    "renderers": {},
-}
-
-ARCHIVE_UI_SCHEMA = {
-    "type": "VerticalLayout",
-    "elements": [
-        {"type": "Control", "scope": "#/properties/name", "customComponent": "ps-input-text"},
-        {
-            "type": "Control",
-            "scope": "#/properties/type",
-            "customComponent": "ps-select",
-        },
-        {
-            "type": "Group",
-            "rule": {
-                "effect": "SHOW",
-                "condition": {"scope": "#/properties/type", "schema": {"const": ""}},
-            },
-            "elements": [],
-        },
-        {"type": "Control", "scope": "#/properties/label", "customComponent": "ps-input-text"},
-        {"type": "Control", "scope": "#/properties/ttl", "customComponent": "ps-input-text"},
-        {
-            "type": "Control",
-            "scope": "#/properties/transform",
-            "customComponent": "ps-textarea-json",
-        },
-        {"type": "Control", "scope": "#/properties/_meta", "customComponent": "ps-textarea-json"},
-    ],
-}
-
 TEST_SCHEMA = {
     "title": "Schema for creating a new test",
     "type": "object",
@@ -550,576 +46,6 @@ TEST_UI_SCHEMA = {
         },
         {"type": "Control", "scope": "#/properties/_meta", "customComponent": "ps-textarea-json"},
     ],
-}
-
-CONTEXT_SCHEMA = {
-    "title": "Schema for creating a new context",
-    "type": "object",
-    "properties": {
-        "name": {
-            "type": "string",
-            "title": "Context Name",
-            "description": "A string to identify this context",
-        },
-        "context": {
-            "type": "string",
-            "title": "Context Type",
-            "description": "Type of context",
-            "oneOf": [],
-            # "oneOf": [
-            #     {"const": "changefail", "title": "Change Fail"},
-            #     {"const": "changenothing", "title": "Change Nothing"},
-            #     {"const": "linuxnns", "title": "Linux NNS"},
-            #     {"const": "linuxvrf", "title": "Linux VRF"},
-            # ],
-        },
-        # "data": {
-        #     "type": "string",
-        #     "title": "Data",
-        #     "description": "JSON object that specifies archive-specific parameters. Archive objects in pSConfig are taken directly from pScheduler. Eg: _url",
-        # },
-        "_meta": {
-            "type": "object",
-            "title": "Other Meta",
-            "description": "Fill in information such as display-name and display-set as an object",
-        },
-    },
-    "required": [
-        "name",
-        "context",
-        "data",
-    ],
-    "renderers": {},
-}
-
-CONTEXT_UI_SCHEMA = {
-    "type": "VerticalLayout",
-    "elements": [
-        {"type": "Control", "scope": "#/properties/name", "customComponent": "ps-input-text"},
-        {
-            "type": "Control",
-            "scope": "#/properties/context",
-            "customComponent": "ps-select",
-        },
-        # {"type": "Control", "scope": "#/properties/data", "customComponent": "ps-textarea-json"},
-        {"type": "Control", "scope": "#/properties/_meta", "customComponent": "ps-textarea-json"},
-    ],
-}
-
-# TODO: Check if we need scheduled-by property, subtasks property
-TASK_SCHEMA = {
-    "title": "Schema for creating a new task",
-    "type": "object",
-    "properties": {
-        "name": {
-            "type": "string",
-            "title": "Task Name",
-            "description": "A string to identify this task",
-        },
-        # "scheduled-by": {
-        #     "title": "Scheduled By",
-        #     "allOf": [
-        #         {
-        #             "type": "integer",
-        #             "minimum": 0
-        #         }
-        #     ]
-        # },
-        "group": {
-            "type": "string",
-            "title": "Group",
-            "oneOf": [],
-        },
-        "test": {
-            "type": "string",
-            "title": "Test",
-            "oneOf": [],
-        },
-        "schedule": {
-            "type": "string",
-            "title": "Schedule",
-            "oneOf": [],
-        },
-        "disabled": {
-            "type": "boolean",
-            "title": "Disabled",
-            "description": "Check this box if this task is disabled. Set to False by default",
-            "default": False,
-        },
-        "archives": {
-            "type": "array",
-            "title": "Archives",
-            "items": {"oneOf": []},
-        },
-        "tools": {
-            "type": "array",
-            "title": "Tools",
-            "items": {"oneOf": []},
-        },
-        # "subtasks": {
-        #     "type": "array",
-        #     "title": "Subtasks",
-        #     "items": {"oneOf": []},
-        # },
-        "priority": {
-            "type": "integer",
-            "title": "Priority",
-        },
-        "reference": {
-            "type": "object",
-            "title": "Reference",
-        },
-        "_meta": {
-            "type": "object",
-            "title": "Other Meta",
-            "description": "Fill in information such as display-name and display-set as an object",
-        },
-    },
-    "required": ["name", "group", "test"],
-    "renderers": {},
-}
-
-TASK_UI_SCHEMA = {
-    "type": "VerticalLayout",
-    "elements": [
-        {
-            "type": "HorizontalLayout",
-            "elements": [
-                {
-                    "type": "Control",
-                    "scope": "#/properties/name",
-                    "customComponent": "ps-input-text",
-                },
-                {
-                    "type": "Control",
-                    "scope": "#/properties/disabled",
-                    "customComponent": "ps-input-checkbox",
-                },
-            ],
-        },
-        {
-            "type": "Control",
-            "scope": "#/properties/group",
-            "customComponent": "ps-select-chip",
-        },
-        {
-            "type": "Control",
-            "scope": "#/properties/test",
-            "customComponent": "ps-select-chip",
-        },
-        {
-            "type": "Control",
-            "scope": "#/properties/archives",
-            "customComponent": "ps-select-multi",
-        },
-        {
-            "type": "Control",
-            "scope": "#/properties/schedule",
-            "customComponent": "ps-select-chip",
-        },
-        {
-            "type": "Control",
-            "scope": "#/properties/tools",
-            "options": {"format": "select"},
-            "customComponent": "ps-select-multi",
-        },
-        {
-            "type": "HorizontalLayout",
-            "elements": [
-                # {
-                #     "type": "Control",
-                #     "scope": "#/properties/tools",
-                #     "options": {"format": "select"},
-                #     "customComponent": "ps-select-multi",
-                # },
-                {
-                    "type": "Control",
-                    "scope": "#/properties/priority",
-                    "customComponent": "ps-input-number",
-                },
-            ],
-        },
-        {
-            "type": "Control",
-            "scope": "#/properties/reference",
-            "customComponent": "ps-textarea-json",
-        },
-        {"type": "Control", "scope": "#/properties/_meta", "customComponent": "ps-textarea-json"},
-    ],
-}
-
-TEMPLATE_SCHEMA = {
-    "title": "Schema for creating a new template",
-    "type": "object",
-    "properties": {
-        "name": {
-            "type": "string",
-            "title": "Name",
-            "description": "A string to identify this template",
-        },
-        "tasks": {
-            "type": "array",
-            "title": "Tasks",
-            "items": {"oneOf": []},
-        },
-        "_meta": {
-            "type": "object",
-            "title": "Other Meta",
-        },
-    },
-    "required": ["name", "tasks"],
-    "renderers": {},
-}
-
-TEMPLATE_UI_SCHEMA = {
-    "type": "VerticalLayout",
-    "elements": [
-        {"type": "Control", "scope": "#/properties/name", "customComponent": "ps-input-text"},
-        {
-            "type": "Control",
-            "scope": "#/properties/tasks",
-            "customComponent": "ps-select-multi",
-        },
-        {"type": "Control", "scope": "#/properties/_meta", "customComponent": "ps-textarea-json"},
-    ],
-}
-
-http_schema = {
-    "json_schema": ARCHIVE_SCHEMA,
-    "ui_schema": ARCHIVE_UI_SCHEMA,
-    "schema": 1,
-    "name": "http",
-    "description": "Send a raw JSON result to a HTTP server",
-    "version": "1.0",
-    "maintainer": {
-        "name": "perfSONAR Development Team",
-        "email": "perfsonar-developer@internet2.edu",
-        "href": "http://www.perfsonar.net",
-    },
-    "spec": {
-        "jsonschema": {
-            "versions": {
-                1: {
-                    "type": "object",
-                    "properties": {
-                        "schema": {"title": "Schema Version", "type": "integer", "enum": [1]},
-                        "_url": {
-                            "title": "URL",
-                            "type": "string",
-                            "format": "uri",
-                        },
-                        "op": {
-                            "title": "Operation",
-                            "type": "string",
-                            "enum": [
-                                "put",
-                                "post",
-                            ],
-                        },
-                        "bind": {
-                            "type": "string",
-                            "minLength": 1,
-                            "maxLength": 255,
-                            "pattern": r"^[A-Za-z0-9_][A-Za-z0-9\-]{0,62}(\.[A-Za-z0-9][A-Za-z0-9\-]{1,62})*\.?$",  # noqa: E501
-                        },
-                        # "retry-policy": {
-                        #     "type": "array",
-                        #     "items": {
-                        #         "$ref": "#/pScheduler/RetryPolicyEntry"
-                        #     }
-                        # }
-                    },
-                    "required": ["_url"],
-                    "additionalProperties": False,
-                },
-                2: {
-                    "type": "object",
-                    "properties": {
-                        "schema": {"title": "Schema Version", "type": "integer", "enum": [2]},
-                        "_url": {
-                            "title": "URL",
-                            "type": "string",
-                            "format": "uri",
-                        },
-                        "op": {
-                            "title": "Operation",
-                            "type": "string",
-                            "enum": [
-                                "put",
-                                "post",
-                            ],
-                        },
-                        "_headers": {
-                            "title": "Headers",
-                            "type": "object",
-                        },
-                        # Not complete
-                        "bind": {
-                            "type": "string",
-                            "minLength": 1,
-                            "maxLength": 255,
-                            "pattern": r"^[A-Za-z0-9_][A-Za-z0-9\-]{0,62}(\.[A-Za-z0-9][A-Za-z0-9\-]{1,62})*\.?$",  # noqa: E501
-                        },
-                        # "retry-policy": {
-                        #     "type": "array",
-                        #     "items": {
-                        #         "$ref": "#/pScheduler/RetryPolicyEntry"
-                        #     }
-                        # }
-                    },
-                    "required": ["schema", "_url"],
-                    "additionalProperties": False,
-                },
-                3: {
-                    "type": "object",
-                    "properties": {
-                        "schema": {"title": "Schema Version", "type": "integer", "enum": [3]},
-                        "_url": {
-                            "title": "URL",
-                            "type": "string",
-                            "format": "uri",
-                        },
-                        "op": {
-                            "title": "Operation",
-                            "type": "string",
-                            "enum": [
-                                "put",
-                                "post",
-                            ],
-                        },
-                        "verify-ssl": {
-                            "title": "Verify SSL Certificate",
-                            "type": "boolean",
-                            "default": False,
-                        },
-                        "_headers": {
-                            "type": "object",
-                        },
-                        # Not complete
-                        "bind": {
-                            "type": "string",
-                            "minLength": 1,
-                            "maxLength": 255,
-                            "pattern": r"^[A-Za-z0-9_][A-Za-z0-9\-]{0,62}(\.[A-Za-z0-9][A-Za-z0-9\-]{1,62})*\.?$",  # noqa: E501
-                        },
-                        # "retry-policy": {
-                        #     "type": "array",
-                        #     "items": {
-                        #         "$ref": "#/pScheduler/RetryPolicyEntry"
-                        #     }
-                        # }
-                    },
-                    "required": ["schema", "_url"],
-                    "additionalProperties": False,
-                },
-                4: {
-                    "type": "object",
-                    "properties": {
-                        "schema": {"title": "Schema Version", "type": "integer", "enum": [4]},
-                        "_url": {
-                            "title": "URL",
-                            "type": "string",
-                            "format": "uri",
-                        },
-                        "op": {
-                            "title": "Operation",
-                            "type": "string",
-                            "enum": [
-                                "put",
-                                "post",
-                            ],
-                        },
-                        "verify-ssl": {
-                            "title": "Verify SSL Certificate",
-                            "type": "boolean",
-                            "default": False,
-                        },
-                        "_headers": {
-                            "title": "Headers",
-                            "type": "object",
-                        },
-                        "timeout": {
-                            "title": "Timeout",
-                            "type": "string",
-                            "pattern": r"^P(?:\d+(?:\.\d+)?W)?(?:\d+(?:\.\d+)?D)?(?:T(?:\d+(?:\.\d+)?H)?(?:\d+(?:\.\d+)?M)?(?:\d+(?:\.\d+)?S)?)?$",  # noqa: E501
-                            "x-invalid-message": "'%s' is not a valid ISO 8601 duration.",
-                        },
-                        # Not complete
-                        "bind": {
-                            "title": "Bind Address",
-                            "type": "string",
-                            "minLength": 1,
-                            "maxLength": 255,
-                            "pattern": r"^[A-Za-z0-9_][A-Za-z0-9\-]{0,62}(\.[A-Za-z0-9][A-Za-z0-9\-]{1,62})*\.?$",  # noqa: E501
-                        },
-                        # "retry-policy": {
-                        #     "type": "array",
-                        #     "items": {
-                        #         "$ref": "#/pScheduler/RetryPolicyEntry"
-                        #     }
-                        # }
-                    },
-                    "required": ["schema", "_url"],
-                    "additionalProperties": False,
-                },
-            },
-        },
-        "uischema": {
-            "versions": {
-                1: {
-                    "type": "VerticalLayout",
-                    "elements": [
-                        {
-                            "type": "Control",
-                            "scope": "#/properties/schema",
-                            "customComponent": "ps-select",
-                        },
-                        {
-                            "type": "Control",
-                            "scope": "#/properties/_url",
-                            "customComponent": "ps-input-text",
-                        },
-                        {
-                            "type": "Control",
-                            "scope": "#/properties/op",
-                            "customComponent": "ps-select",
-                        },
-                        {
-                            "type": "Control",
-                            "scope": "#/properties/bind",
-                            "customComponent": "ps-input-text",
-                        },
-                        # {
-                        #     "type": "Control",
-                        #     "scope": "#/properties/retry-policy",
-                        #     "customComponent": "ps-input-text",
-                        # },
-                    ],
-                },
-                2: {
-                    "type": "VerticalLayout",
-                    "elements": [
-                        {
-                            "type": "Control",
-                            "scope": "#/properties/schema",
-                            "customComponent": "ps-select",
-                        },
-                        {
-                            "type": "Control",
-                            "scope": "#/properties/_url",
-                            "customComponent": "ps-input-text",
-                        },
-                        {
-                            "type": "Control",
-                            "scope": "#/properties/op",
-                            "customComponent": "ps-select",
-                        },
-                        {
-                            "type": "Control",
-                            "scope": "#/properties/_headers",
-                            "customComponent": "ps-textarea-json",
-                        },
-                        {
-                            "type": "Control",
-                            "scope": "#/properties/bind",
-                            "customComponent": "ps-input-text",
-                        },
-                        # {
-                        #     "type": "Control",
-                        #     "scope": "#/properties/retry-policy",
-                        #     "customComponent": "ps-input-text",
-                        # },
-                    ],
-                },
-                3: {
-                    "type": "VerticalLayout",
-                    "elements": [
-                        {
-                            "type": "Control",
-                            "scope": "#/properties/schema",
-                            "customComponent": "ps-select",
-                        },
-                        {
-                            "type": "Control",
-                            "scope": "#/properties/_url",
-                            "customComponent": "ps-input-text",
-                        },
-                        {
-                            "type": "Control",
-                            "scope": "#/properties/op",
-                            "customComponent": "ps-select",
-                        },
-                        {
-                            "type": "Control",
-                            "scope": "#/properties/verify-ssl",
-                            "customComponent": "ps-input-checkbox",
-                        },
-                        {
-                            "type": "Control",
-                            "scope": "#/properties/_headers",
-                            "customComponent": "ps-textarea-json",
-                        },
-                        {
-                            "type": "Control",
-                            "scope": "#/properties/bind",
-                            "customComponent": "ps-input-text",
-                        },
-                        # {
-                        #     "type": "Control",
-                        #     "scope": "#/properties/retry-policy",
-                        #     "customComponent": "ps-input-text",
-                        # },
-                    ],
-                },
-                4: {
-                    "type": "VerticalLayout",
-                    "elements": [
-                        {
-                            "type": "Control",
-                            "scope": "#/properties/schema",
-                            "customComponent": "ps-select",
-                        },
-                        {
-                            "type": "Control",
-                            "scope": "#/properties/_url",
-                            "customComponent": "ps-input-text",
-                        },
-                        {
-                            "type": "Control",
-                            "scope": "#/properties/op",
-                            "customComponent": "ps-select",
-                        },
-                        {
-                            "type": "Control",
-                            "scope": "#/properties/_headers",
-                            "customComponent": "ps-textarea-json",
-                        },
-                        {
-                            "type": "Control",
-                            "scope": "#/properties/verify-ssl",
-                            "customComponent": "ps-input-checkbox",
-                        },
-                        {
-                            "type": "Control",
-                            "scope": "#/properties/timeout",
-                            "customComponent": "ps-input-text",
-                        },
-                        {
-                            "type": "Control",
-                            "scope": "#/properties/bind",
-                            "customComponent": "ps-input-text",
-                        },
-                        # {
-                        #     "type": "Control",
-                        #     "scope": "#/properties/retry-policy",
-                        #     "customComponent": "ps-input-text",
-                        # },
-                    ],
-                },
-            },
-        },
-        "versions": [None, "1", "2", "3", "4"],
-    },
 }
 
 idle_schema = {
@@ -2462,159 +1388,4 @@ latency_schema = {
         },
         "versions": [None, "3", "4"],
     },
-}
-
-TEMPLATE_IMPORT_SCHEMA = {
-    "title": "Schema for creating a new template via import",
-    "type": "object",
-    "properties": {
-        "name": {
-            "type": "string",
-            "title": "Name",
-            "description": "A name to identify this template.",
-        },
-        "resolveConflict": {
-            "type": "string",
-            "title": "Conflict Resolution Strategy",
-            "oneOf": [
-                {"const": "overwrite", "title": "Overwrite (Keep New)"},
-                {"const": "keep-existing", "title": "Keep Existing"},
-                {"const": "keep-both", "title": "Keep Both (Auto Rename)"},
-            ],
-            "description": "Choose how to resolve datatypes with same name",
-        },
-        "orphanData": {
-            "type": "boolean",
-            "title": "Import Orphan Datas",
-            "description": "Any child data types that aren’t referenced in the parent template will be created if checked.",
-        },
-        "importVia": {
-            "type": "string",
-            "title": "Import Via",
-            "oneOf": [
-                {"const": "upload", "title": "Upload JSON File"},
-                {"const": "url", "title": "Import from URL"},
-                {"const": "paste", "title": "Paste JSON"},
-            ],
-        },
-        "upload": {
-            "type": "string",
-            "title": "Upload JSON File",
-        },
-        "url": {
-            "type": "string",
-            "title": "URL",
-        },
-        "paste": {
-            "type": "object",
-            "title": "Paste JSON",
-        },
-        "_meta": {
-            "type": "object",
-            "title": "Other Meta",
-        },
-    },
-    "required": ["name", "importVia"],
-    "allOf": [
-        {
-            "if": {"properties": {"importVia": {"const": "upload"}}},
-            "then": {"required": ["upload"]},
-        },
-        {
-            "if": {"properties": {"importVia": {"const": "url"}}},
-            "then": {"required": ["url"]},
-        },
-        {
-            "if": {"properties": {"importVia": {"const": "paste"}}},
-            "then": {"required": ["paste"]},
-        },
-    ],
-}
-
-
-TEMPLATE_IMPORT_UI_SCHEMA = {
-    "type": "VerticalLayout",
-    "elements": [
-        {
-            "type": "Control",
-            "scope": "#/properties/name",
-            "customComponent": "ps-input-text",
-        },
-        {
-            "type": "HorizontalLayout",
-            "elements": [
-                {
-                    "type": "Control",
-                    "scope": "#/properties/resolveConflict",
-                    "customComponent": "ps-select",
-                },
-                {
-                    "type": "Control",
-                    "scope": "#/properties/orphanData",
-                    "customComponent": "ps-input-checkbox",
-                },
-            ],
-        },
-        {
-            "type": "Control",
-            "scope": "#/properties/importVia",
-            "customComponent": "ps-select",
-        },
-        {
-            "type": "Template",
-            "rule": {
-                "effect": "SHOW",
-                "condition": {
-                    "scope": "#/properties/importVia",
-                    "schema": {"const": "upload"},
-                },
-            },
-            "elements": [
-                {
-                    "type": "Control",
-                    "scope": "#/properties/upload",
-                    "customComponent": "ps-input-file",
-                },
-            ],
-        },
-        {
-            "type": "Template",
-            "rule": {
-                "effect": "SHOW",
-                "condition": {
-                    "scope": "#/properties/importVia",
-                    "schema": {"const": "url"},
-                },
-            },
-            "elements": [
-                {
-                    "type": "Control",
-                    "scope": "#/properties/url",
-                    "customComponent": "ps-input-text",
-                },
-            ],
-        },
-        {
-            "type": "Template",
-            "rule": {
-                "effect": "SHOW",
-                "condition": {
-                    "scope": "#/properties/importVia",
-                    "schema": {"const": "paste"},
-                },
-            },
-            "elements": [
-                {
-                    "type": "Control",
-                    "scope": "#/properties/paste",
-                    "customComponent": "ps-textarea-json",
-                },
-            ],
-        },
-        {
-            "type": "Control",
-            "scope": "#/properties/_meta",
-            "customComponent": "ps-textarea-json",
-        },
-    ],
 }
